+++ workspace/node/bookmarks/tests/parsers/xbel.test.ts
@@ -1,6 +1,5 @@
 import {
     readFile,
-    // writeFile,
 } from "node:fs/promises";
 
 import {