+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 更改日志 | Change Log
 
+## 2025-01-17
+
+- 实现 `XBEL` 格式的书签构造器 | Implement a builder for bookmarks in `XBEL` format
+
 ## 2025-01-12
 
 - 修复 `XBEL` 解析器无法解析 `desc` 字段的问题 | Fix the issue that the `XBEL` parser cannot parse the `desc` field